requirements.txt
feedparser>=6.0
dnspython>=2.4
requests>=2.31
flask>=3.0
openai>=1.0
anthropic>=0.30
python-telegram-bot>=21.0
beautifulsoup4>=4.12
pyyaml>=6.0
python-dotenv>=1.0
